+++ apps/app/src/server/service/interfaces/export.ts
@@ -1,0 +1,13 @@
+import { Stats } from 'fs';
+
+export type ZipFileStat = {
+  meta: object;
+  fileName: string;
+  zipFilePath: string;
+  fileStat: Stats;
+  innerFileStats: {
+      fileName: string;
+      collectionName: string;
+      size: number;
+  }[];
+} | null
